Zmiana funkcji obrotu

+++ domena/sense_hat_controller.py
@@ -30,7 +30,7 @@ class SenseHatController(SenseHat):
         # kat nie moze byc wiekszy od 360
         self.__kat = int(self.__kat) % 360
         # zaleznie od kata 0 -360 wybieramy odpowiedni stan od 1 do 20
-        self.__stan = self.__kat / 18 + 1
+        self.__stan = int(self.__kat / 18) + 1
 
         self.__silnik_obrot = STANY[self.__stan]
         self.przygotowanie_silnika_do_wyswietlenia()
